Review and changes /10-02/requirements.txt
# CuKEM: Hybrid PQC-QKD System Requirements

# Core cryptography
cryptography>=41.0.0
liboqs-python>=0.8.0

# Quantum simulation
qiskit>=0.45.0
qiskit-aer>=0.14.0
qiskit-ibm-runtime>=0.18.0

# Configuration management
PyYAML>=6.0.0

# State machine
transitions>=0.9.0

# Numerical and scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and monitoring
python-json-logger>=2.0.7

# Network simulation (optional - Linux only)
# mininet>=2.3.0

# Utilities
python-dotenv>=1.0.0
